+++ CHANGELOG.md
@@ -31,6 +31,7 @@ BUG FIXES:
       WinRM in some cases. [GH-3816]
   - hosts/windows: Don't execute mstsc using PowerShell since it doesn't
       exit properly. [GH-3837]
+  - hosts/windows: For RDP, don't remove the Tempfile right away. [GH-3875]
   - providers/docker: Never do graceful shutdown, always use
       `docker stop`. [GH-3798]
   - providers/docker: Better error messaging when SSH is not ready